# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel 8 OpenApi TV Channel Documentation",
        "description": "Laravel 8 OpenApi TV Channel description, Cross-platform lab 1-6",
        "contact": {
            "email": "kilubaka@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/programs": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Get list of programs",
                "description": "Returns list of programs",
                "operationId": "getProgramsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Programs"
                ],
                "summary": "Store new program",
                "description": "Returns created program data",
                "operationId": "storeProgram",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProgramRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/programs/{id}": {
            "get": {
                "tags": [
                    "Programs"
                ],
                "summary": "Get program information",
                "description": "Returns program data",
                "operationId": "getProgramById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Program id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Programs"
                ],
                "summary": "Update existing program",
                "description": "Returns updated program data",
                "operationId": "updateProgram",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Program id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProgramRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Programs"
                ],
                "summary": "Delete existing program",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteProgram",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Program id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/schedules": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get list of schedules",
                "description": "Returns list of schedules",
                "operationId": "getSchedulesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Store new schedule",
                "description": "Returns created schedule data",
                "operationId": "storeSchedule",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreScheduleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/schedules/{id}": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get schedule information",
                "description": "Returns schedule data",
                "operationId": "getScheduleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Schedule id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Update existing schedule",
                "description": "Returns updated schedule data",
                "operationId": "updateSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Schedule id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateScheduleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Delete existing schedule",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Schedule id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreProgramRequest": {
                "title": "Store Program request",
                "description": "Store Program request body data",
                "required": [
                    "name",
                    "description",
                    "category_id",
                    "age_restriction",
                    "cost"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new program",
                        "type": "string",
                        "format": "string",
                        "example": "Nostrum voluptas."
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the Program",
                        "type": "string",
                        "format": "string",
                        "example": "Officiis odit ratione quos aspernatur."
                    },
                    "category_id": {
                        "title": "Category ID",
                        "description": "Category ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "age_restriction": {
                        "title": "Age Restriction",
                        "description": "Age Restriction",
                        "type": "integer",
                        "format": "int64",
                        "example": 18
                    },
                    "cost": {
                        "title": "Cost",
                        "description": "Cost",
                        "type": "number",
                        "format": "float",
                        "example": 1200.5
                    }
                },
                "type": "object"
            },
            "StoreScheduleRequest": {
                "title": "Store Schedule request",
                "description": "Store Schedule request body data",
                "required": [
                    "from",
                    "to",
                    "program_id",
                    "leader_id"
                ],
                "properties": {
                    "from": {
                        "title": "From",
                        "description": "Datetime from",
                        "type": "string",
                        "format": "date-time",
                        "example": "2017-07-21 17:32:28"
                    },
                    "to": {
                        "title": "To",
                        "description": "Datetime to",
                        "type": "string",
                        "format": "date-time",
                        "example": "2017-07-21 17:32:28"
                    },
                    "program_id": {
                        "title": "Program ID",
                        "description": "Program ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "leader_id": {
                        "title": "Leader ID",
                        "description": "Leader ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "UpdateProgramRequest": {
                "title": "Update Program request",
                "description": "Update Program request body data",
                "required": [
                    "name",
                    "description",
                    "category_id",
                    "age_restriction",
                    "cost"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the update program",
                        "type": "string",
                        "format": "string",
                        "example": "Nostrum voluptas."
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the Program",
                        "type": "string",
                        "format": "string",
                        "example": "Officiis odit ratione quos aspernatur."
                    },
                    "category_id": {
                        "title": "Category ID",
                        "description": "Category ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "age_restriction": {
                        "title": "Age Restriction",
                        "description": "Age Restriction",
                        "type": "integer",
                        "format": "int64",
                        "example": 18
                    },
                    "cost": {
                        "title": "Cost",
                        "description": "Cost",
                        "type": "number",
                        "format": "float",
                        "example": 1200.5
                    }
                },
                "type": "object"
            },
            "UpdateScheduleRequest": {
                "title": "Update Schedule request",
                "description": "Update Schedule request body data",
                "required": [
                    "from",
                    "to",
                    "program_id",
                    "leader_id"
                ],
                "properties": {
                    "from": {
                        "title": "From",
                        "description": "Datetime from",
                        "type": "string",
                        "format": "date-time",
                        "example": "2017-07-21 17:32:28"
                    },
                    "to": {
                        "title": "To",
                        "description": "Datetime to",
                        "type": "string",
                        "format": "date-time",
                        "example": "2017-07-21 17:32:28"
                    },
                    "program_id": {
                        "title": "Program ID",
                        "description": "Program ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "leader_id": {
                        "title": "Leader ID",
                        "description": "Leader ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}